Rearranged folder structure, made separate apps within same project for different operations
