+++ .travis.yml
@@ -7,7 +7,7 @@ python:
 
 before_install:
   - sudo apt-get install libmysqlclient-dev
-  - cd mbtb_app/resources/apis/user_registration
+  - cd mbtb_app/resources/apis/users
   - pip install -r requirements.txt
 
 script:
