--- mbtb_app/resources/apis/user_registration/register/serializers.py
@@ -1,16 +1,0 @@
-from rest_framework import serializers
-from .models import AdminAccount, Users
-
-
-class AdminAuthenticationSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = AdminAccount
-        fields = '__all__'
-
-
-class UsersSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Users
-        fields = '__all__'
