+++ mbtb_app/resources/apis/users/admin_signin/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AdminSigninConfig(AppConfig):
+    name = 'admin_signin'
