+++ mbtb_app/resources/apis/users/admin_signin/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path, include
+from . import views
+from rest_framework import routers
+
+router = routers.DefaultRouter()
+
+urlpatterns = [
+    path('', include(router.urls)),
+    path('admin_auth', views.AdminAccountView.as_view())
+]
