+++ mbtb_app/resources/apis/users/register/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
