+++ mbtb_app/resources/apis/users/register/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from .models import Users
+
+
+class UsersSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = Users
+        fields = '__all__'
