+++ mbtb_app/resources/apis/users/register/urls.py
@@ -4,9 +4,8 @@ from rest_framework import routers
 
 router = routers.DefaultRouter()
 router.register('list_new_users', views.NewUsersListViewSet)
-router.register('new_users', views.NewUsersViewSet)
+router.register('add_new_users', views.NewUsersViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('admin_auth', views.AdminAccountView.as_view())
 ]
