+++ mbtb_app/resources/apis/users/users/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for user_registration project.
+Django settings for users project.
 
 Generated by 'django-admin startproject' using Django 2.2.5.
 
@@ -41,6 +41,8 @@ INSTALLED_APPS = [
     'rest_framework',
     'rest_framework.authtoken',
     'register',
+    'admin_signin',
+    'users_signin',
 ]
 
 MIDDLEWARE = [
@@ -53,7 +55,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'user_registration.urls'
+ROOT_URLCONF = 'users.urls'
 
 TEMPLATES = [
     {
@@ -71,7 +73,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'user_registration.wsgi.application'
+WSGI_APPLICATION = 'users.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/2.2/ref/settings/#databases
