+++ mbtb_app/resources/apis/users/users/urls.py
@@ -1,4 +1,4 @@
-"""user_registration URL Configuration
+"""users URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.2/topics/http/urls/
@@ -18,5 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('register.urls'))
+    path('', include('register.urls')),
+    path('', include('admin_signin.urls')),
 ]
