+++ mbtb_app/resources/apis/users/users_signin/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UsersSigninConfig(AppConfig):
+    name = 'users_signin'
